refactor: Convert background script to typescript
